+++ src/PSRest/Const.cs
@@ -1,5 +1,4 @@
-﻿
-namespace PSRest;
+﻿namespace PSRest;
 
 static class Const
 {
@@ -9,11 +8,20 @@ static class Const
     public const string VarRestEnvironment = "RestEnvironment";
     public const string VSCodeDir = ".vscode";
 
+    public const string DatetimeRfc1123 = "rfc1123";
+    public const string DatetimeIso8601 = "iso8601";
+
     public const string EnvNameShared = "$shared";
     public const string EnvVarDefault = "REST_ENV";
 
     public const string MediaTypeJson = "application/json";
     public const string MediaTypeXml = "application/xml";
+
+    public const string SystemGuid = "$guid";
+    public const string SystemRandomInt = "$randomInt";
+    public const string SystemTimestamp = "$timestamp";
+    public const string SystemDatetime = "$datetime";
+    public const string SystemLocalDatetime = "$localDatetime";
 
     public const string VarTypeShared = "$shared";
     public const string VarTypeDotEnv = "$dotenv";